Remove config and resources singleton logic
The global objects for keeping configuration and resources are not
created as singletons anymore, but are created at the beginning of
the test run and kept in the RSpec configuration as custom attributes.
This makes the CFValidator class redundant. They are accessed as
Validator::Api.configuration and Validator::Api.resources now.

Signed-off-by: Tom Kiemes <tom.kiemes@sap.com>