Check that there is no remote security group configuered.
Security rules can have a remoted_security group, that limits access
to instances with the configured security group. The default for the
'default' security group is to have an ingres rule, that allows acces
for all tcp protocols, however limited to the default security group.
Access from outside is not allowed. However the test for ssh access
did not fail.