Explicitely list/filter entities by project
The script is dangerous. If the user running the CFOV has an admin role
on the project the validator runs in, the cleanup script will find all
glance images and all neutron ports.
The list commands filters by default on the project for servers, volumes
and snapshot.
I added the project filter explicitely for servers, images and ports.
(For volumes there is also a `--project` option, but passing this did
not worked for me with the cleanup script.)

And yes openstack CLIs are great as they are everything else then
consistent across projects :)