Make swift directory cleanup more gracefully
Due to the ordering of clean up in the resource tracker it could happen
that files where deleted before directories. Through the distributed
nature of blobstore nodes it was possible that a validator test failed
because it tried to delete a file inside a directory which was
already deleted on another node.

[#158837753](https://www.pivotaltracker.com/story/show/158837753)

Co-authored-by: Max Becker <max.becker@sap.com>